% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/patterns.R
\name{apply.pattern}
\alias{apply.pattern}
\title{Apply Variables to File Pattern}
\usage{
apply.pattern(name, ...)
}
\arguments{
\item{...}{variables to insert}
}
\value{
character vector with variables inserted where needed
}
\description{
User passes a pattern to use and the variable/value pairs to insert where specified in the pattern.
For example, if you store your annual data under the data folder by year, you might have a pattern named
'annual_data_folder' that is './data/[YEAR]/' and you would call
apply.pattern("annual_data_folder", YEAR = 2020)
}
\details{
There is one special case. If you pass YEAR as a variable, it will also create YR, and the
correesponding 2-digit value.
}
\examples{
\dontrun{
sasout_file<-apply.pattern("sas_sasout_version",YEAR = year, VERS = version)

}

}
