% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ascii_data.R
\name{ascii.download.data}
\alias{ascii.download.data}
\title{Download ASCII BRFSS data}
\usage{
ascii.download.data(
  destpath = NULL,
  unzip = TRUE,
  rmzip = TRUE,
  progress = NULL
)
}
\arguments{
\item{destpath}{character - location to store file - will be created if it does not exist - default is the data_raw folder in the project root folder}

\item{unzip}{logical - unzip file?}

\item{rmzip}{logical - remove the zip file?}

\item{year}{integer - year of data file}
}
\description{
Download the annual ASCII BRFSS data files, all versions
}
\examples{
\dontrun{
ascii.download.data(2016)
}


}
