% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/survey.R
\name{binary_no}
\alias{binary_no}
\title{Value Representing Binary No}
\usage{
binary_no()
}
\value{
integer - value representing 'No'
}
\description{
Retrieve the value used to represent 'No' in binary (Yes/No) calculations.
When there are more than 2 possible values in the denominator and/or numerator,
then it is convenient to convert all values in a numerator to Yes/No
}
