% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/survey.R
\name{binary_yes}
\alias{binary_yes}
\title{Value Representing Binary Yes}
\usage{
binary_yes()
}
\value{
integer - value representing 'Yes'
}
\description{
Retrieve the value used to represent 'Yes' in binary (Yes/No) calculations.
When there are more than 2 possible values in the denominator and/or numerator,
then it is convenient to convert all values in a numerator to Yes/No
}
