% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read.brfss.R
\name{brfss.raw.ascii.filename}
\alias{brfss.raw.ascii.filename}
\title{Build BRFSS raw data file (ASCII) filename}
\usage{
brfss.raw.ascii.filename(year, month, index = 1, type, state = "VI",
  ynpath = F)
}
\arguments{
\item{year}{integer year of the data included in the file}

\item{month}{integer month of the data included in the file}

\item{index}{integer index of the raw data file}

\item{type}{character either cell or landline}

\item{state}{character include data for "all" states (default), "my" state, or "other" states}
}
\value{
character filename based on passed arguments
}
\description{
Build BRFSS raw data file filename based on the year, month, index, type, and state
}

