% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/survey.R
\name{brfss_data}
\alias{brfss_data}
\title{Get BRFSS Survey Data}
\usage{
brfss_data(...)
}
\arguments{
\item{year}{\itemize{
\item int - year of interest
}}

\item{geog}{\itemize{
\item character - 2-char geog of interest (ex: "NY")
}}

\item{version}{\itemize{
\item survey version; 0 (default) = main survey
}}

\item{extent}{\itemize{
\item character - extent of data ("local" or public")
}}

\item{source}{\itemize{
\item character - source of data ("ascii" or sas")
}}
}
\value{
data frame
}
\description{
Get BRFSS Survey Data
}
\examples{
\dontrun{
df <- brfss_data(2020, geog = "MD", version = 1)

}

}
