% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/brfss_paths.R
\name{brfss_data_path}
\alias{brfss_data_path}
\title{Get BRFSS data file path}
\usage{
brfss_data_path(rw = c("r", "w"))
}
\arguments{
\item{rw}{character - read/write}

\item{year}{integer - year of interest}

\item{geog}{character - geography of interest}

\item{version}{integer - survey version (0 = main)}

\item{extent}{character -'local' or 'national'}

\item{source}{character - data source ('ascii' or 'sas')}
}
\description{
This is used in two ways based on the `rw` variable.
}
\details{
1. rw = 'w': we want to write it ... find the path and create the folder so the file can be saved

2. rw = 'r': we want to read it ... find the path of the file to be read and return NULL if it doesn't exist
}
