% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/brfss_paths.R
\name{brfss_file_info}
\alias{brfss_file_info}
\title{BRFSS Data File Information}
\usage{
brfss_file_info(filename)
}
\arguments{
\item{filename}{character name of the file, may include the path}
}
\value{
list containing members path, type, year, month, index
}
\description{
Gets information for a filename containing BRFSS raw data
}
\examples{
\dontrun{
filename<-"./data_raw/2016/04_APR/CEL_VIAPR161.dat"
lst<-brfss_file_info(filename)
year<-lst$year
}
}
