% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/brfss_paths.R
\name{brfss_layout_filename}
\alias{brfss_layout_filename}
\title{BRFSS data file processing}
\usage{
brfss_layout_filename(year)
}
\arguments{
\item{year}{integer year of the data included in the file}
}
\value{
character default path/filename
}
\description{
Gets default path and filename for a file containing layout data for a BRFSS raw data file
}

