% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/brfss_paths.R
\name{brfss_raw_filename}
\alias{brfss_raw_filename}
\title{BRFSS data file default name}
\usage{
brfss_raw_filename(
  type = c("landline", "cell"),
  year,
  month,
  index = 1,
  ynpath = T
)
}
\arguments{
\item{type}{character  source of the data in the file ... "landline" or "cell"}

\item{year}{integer year of the data included in the file}

\item{month}{integer month of the data included in the file}

\item{index}{integer index of the raw data file}

\item{ynpath}{logical (default=TRUE) include the default path?}
}
\value{
character default path
}
\description{
Gets default name for a file containing BRFSS raw data, including the path if requested
}
\examples{
\dontrun{
filename_raw<-brfss_raw_filename("cell",2016,2)
}

}
