% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/brfss_paths.R
\name{brfss_raw_path}
\alias{brfss_raw_path}
\title{BRFSS data file default path}
\usage{
brfss_raw_path(year, month)
}
\arguments{
\item{year}{integer year of the data included in the file}

\item{month}{integer month of the data included in the file}
}
\value{
character default path
}
\description{
Gets default path for a file containing BRFSS raw data
}
\examples{
\dontrun{
path<-brfss_raw_path(2016,2)
}

}

