% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/saq.R
\name{build_saq_layout}
\alias{build_saq_layout}
\title{Build State-Added-Questions Layout File}
\usage{
build_saq_layout()
}
\description{
Read in .csv file and save as .rda file
}
\examples{
\dontrun{
build_saq_layout(2021)
}
}
