% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/report_funcs.R
\name{cleanDesc}
\alias{cleanDesc}
\title{BRFSS data file processing}
\usage{
cleanDesc(desc)
}
\arguments{
\item{desc}{character string to be cleaned}
}
\value{
character cleaned string vector
}
\description{
Clean the description field
}

