% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ascii_data.R
\name{cleave.geogs.ascii}
\alias{cleave.geogs.ascii}
\title{Split BRFSS ASCII data by Geography}
\usage{
cleave.geogs.ascii(
  year = NULL,
  main = TRUE,
  versions = TRUE,
  my_geog = NULL,
  other_geogs = NULL,
  verbose = TRUE
)
}
\arguments{
\item{year}{integer - year of interest}

\item{main}{logical - process main ascii data file}

\item{versions}{logical - process versioned ascii file}

\item{my_geog}{character - abbreviation for primary state/geography of interest (e. "MT")}

\item{other_geogs}{character - abbreviations for other states/geographies of interest (e. c("ID","WY"))}

\item{verbose}{logical - provide details during processing}
}
\description{
The main BRFSS ASCII data file created when the XPT file is downloaded and read has data
for all geographies. This function splits out the geographies of interest.
}
\examples{

\dontrun{
cleave.geogs.ascii(year = 2020, main=TRUE,versions=TRUE, my_geog="MT", other_geogs=NULL,verbose=TRUE)
}

}
