% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sas.R
\name{cleave.geogs}
\alias{cleave.geogs}
\title{Split BRFSS XPT by Geography}
\usage{
cleave.geogs(
  year = NULL,
  rdata_folder = NULL,
  rdata_file = NULL,
  main = TRUE,
  versions = TRUE,
  my_geog = NULL,
  other_geogs = NULL,
  verbose = TRUE
)
}
\arguments{
\item{year}{integer - year of interest}

\item{rdata_folder}{character - location of .RData file with full BRFSS XPT data}

\item{rdata_file}{character - filename of .RData file with full BRFSS XPT data}

\item{main}{logical - process main XPT file}

\item{versions}{logical - process versioned XPT file}

\item{my_geog}{character - abbreviation for primary state/geography of interest (e. "MT")}

\item{other_geogs}{character - abbreviations for other states/geographies of interest (e. c("ID","WY"))}

\item{verbose}{logical - provide details during processing}
}
\description{
The main BRFSS XPT data file created when the XPT file is downloaded and read has data
for all geographies.This function splits out the geographies of interest.
}
\examples{

\dontrun{
cleave.geogs(year = 2020,rdata_folder="./data/2020/", rdata_file="xpt_2020.RData",
   main=TRUE,versions=TRUE, my_geog="MT", other_geogs=NULL,verbose=TRUE)
}

}
