% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/layout.R
\name{codebook_layout}
\alias{codebook_layout}
\title{Build the Layout from CDC Provided Codebook}
\usage{
codebook_layout(file = NULL)
}
\arguments{
\item{file}{character name of layout file - if missing (NULL) then my.brfss() is called to
determine the year and geographical place of interest. The geographical place of interest
may not be used, but will be passed for pattern matching of file name in case it is needed}
}
\value{
data.frame containing the layout to be passed to read.fwf
}
\description{
If a filename is not provided, it uses the file_pattern data.frame
to figure out the name of the file. The layout data.frame is used to
parse the BRFSS ascii fixed width file
}
