% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/values.R
\name{codebook_values}
\alias{codebook_values}
\title{Codebook Values}
\usage{
codebook_values(file = NULL, year = NULL)
}
\arguments{
\item{file}{character - filename of layout file}

\item{year}{integer - year of interest for codebook}
}
\value{

}
\description{
The annual codebook is parsed for possible values for each variable.
}
