% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ascii_data.R
\name{convert.ascii}
\alias{convert.ascii}
\title{Convert BRFSS raw ASCII data files}
\usage{
convert.ascii(
  year = NULL,
  layout = NULL,
  completes = T,
  main = TRUE,
  versions = TRUE,
  verbose = FALSE
)
}
\arguments{
\item{year}{integer year of interest}

\item{layout}{character name of the file containing the layout or data.frame containing the layout}

\item{completes}{logical whether or not to include only complete interviews (default = TRUE)}

\item{main}{logical include main ascii file (version 0) (default = TRUE)}

\item{versions}{logical include ascii file versions (default = TRUE)}

\item{verbose}{logical print some extra information during processing (default = FALSE)}
}
\value{
data frame containing survey data
}
\description{
Convert BRFSS raw ASCII data files to data frames and save
}
\examples{
\dontrun{
my.brfss(year = 2020, geog = "MT")
df<-read.brfss.ascii(year = ,
layout="./data_raw/var_layouts/layout_2016")
}

}
