% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/codebook.R
\name{download.codebook}
\alias{download.codebook}
\title{Download Codebook}
\usage{
download.codebook(year = NULL, geog = NULL)
}
\arguments{
\item{year}{integer - year of interest}
}
\description{
Download the CDC prepared codebook for the BRFSS data files. The format of the URL is not consistent by year
so more than one pattern is attempted and all should fail but one for a given year.
}
\examples{
\dontrun{
download.codebook(2020)
}
}
