% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ascii_data.R
\name{download_ascii_data}
\alias{download_ascii_data}
\title{Download ASCII BREFSS data}
\usage{
download_ascii_data(year, destpath = "./data_raw/", unzip = TRUE, rmzip = TRUE)
}
\arguments{
\item{rmzip}{}
}
\value{

}
\description{
Download ASCII BREFSS data
}
\examples{
download_ascii_data(2016)

}
