% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/codebook.R
\name{download_codebook}
\alias{download_codebook}
\title{Download Codebook}
\usage{
download_codebook(fldrout = NULL, year = NULL, progress = NULL)
}
\arguments{
\item{fldrout}{character - folder to store codebook - NULL gets default}

\item{year}{integer - year of interest}

\item{progress}{progress object}
}
\description{
Download the CDC prepared codebook for the BRFSS data files. The format of the URL is not consistent by year
so more than one pattern is attempted and all should fail but one for a given year.
}
\examples{
\dontrun{
download_codebook(2020)
}
}
