% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fetch_brfss.R
\name{fetch_brfss}
\alias{fetch_brfss}
\title{Download BRFSS SMART data}
\usage{
fetch_brfss(year = 2016, recode = TRUE, smart = TRUE, filepath = NULL)
}
\arguments{
\item{year}{A year or vector of years between 2002 & 2016.}

\item{recode}{Whether to recode select variables so that variable names are consistent over time.}
}
\value{
A tbl.
}
\description{
`fetch_brfss()` is a convenient wrapper to download BRFSS SMART data.
}
