% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geogs.R
\name{geog_abb}
\alias{geog_abb}
\title{Get Geography Abbreviation}
\usage{
geog_abb(geogs)
}
\arguments{
\item{geog}{integer or character - either the full geography name or the FIPS ID}
}
\value{
character abbreviation of geography (e.g "MT" or "VI")
}
\description{
Get the abbreviation variable for a geography based on the numeric id or the full name
}
\examples{
\dontrun{
# For Montana  ...
geog_abb(30)
# or
geog_abb("Montana")
}

}
