% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geogs.R
\name{geog_id}
\alias{geog_id}
\title{Get Geography Numeric (FIPS) ID}
\usage{
geog_id(geogs)
}
\arguments{
\item{geog}{character - either the full geography name or the 2-character abbreviation}
}
\value{
integer - id of geography
}
\description{
Get the id variable for a geography based on the full name or the abbreviation for the geography.
}
\examples{
\dontrun{
# For Montana  ...
geog_id("MT")
# or
geog_id("Montana")
}

}
