% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modules.R
\name{geog_modules}
\alias{geog_modules}
\title{Get BRFSS Modules Used by State(s)}
\usage{
geog_modules(year = NULL, geogs = NULL, versions = FALSE, reduce = TRUE)
}
\arguments{
\item{year}{- integer - 4-digit year of interest}

\item{geogs}{- character or integer vector - geogs of interest - can be 2 character geog code or FIPS code (integer)}

\item{versions}{- logical - include the version number}

\item{reduce}{- logical - reduce to a vector if appropriate}
}
\value{

}
\description{
Returns  a data frame that has the year, the version (if requested), the geog (2-char), and the module used or returns a
vector of modules if appropriate
}
\examples{
\dontrun{
geog_modules(2018,"WI")
geog_modules(2018,c(1,2))
geog_modules(2018,"MI",versions=T)
}

}
