% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geogs.R
\name{geog_name}
\alias{geog_name}
\title{Get Geography Name}
\usage{
geog_name(geog)
}
\arguments{
\item{geog}{integer or character - either the 2-character geography abbreviation or the FIPS ID}
}
\value{
character name of geography (e.g "Montana" or "Virgin Islands")
}
\description{
Get the name variable for a geography based on the numeric id or the abbreviation.
}
\examples{
\dontrun{
# For Alabama  ...
geog_name(1)
# or
geog_name("AL")
}
}
