% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{geogs}
\alias{geogs}
\title{Storage for Default Year and State/Geography}
\format{
A data frame with 3 variables and 54 rows
\describe{

  \item{Geog}{text name of geography (e.g. "Montana")}
  \item{Id}{BRFSS integer ID for the Geography - equivalent to the FIPS code (e.g. 30)}
  \item{Abbrev}{2-digit abbreviation for the geography (state/territory) (e.g. "MT")}
  ...
}
}
\usage{
geogs
}
\description{
A list containing  information for current BRFSS processing
}
\keyword{datasets}
