% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/layout.R
\name{get.layout}
\alias{get.layout}
\title{Get Layout}
\usage{
get.layout(year = NULL, ...)
}
\arguments{
\item{...}{other arguments to be passed to apply.pattern()}
}
\value{
data.frame containing layout information
}
\description{
First, an attempt is made to get the codebook layout. If that fails, an attempt is made to get the sas layout.
}
