% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/codebook.R
\name{get.merged.layout}
\alias{get.merged.layout}
\title{Get Merged Layout}
\usage{
get.merged.layout()
}
\arguments{
\item{year}{integer - year of interest}
}
\value{
data frame with layout data
}
\description{
Get layout created from merging national layout with state-added questions
}
