% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/patterns.R
\name{get.pattern.group}
\alias{get.pattern.group}
\title{Get File Pattern Group}
\usage{
get.pattern.group(group)
}
\arguments{
\item{group}{character: name of pattern group}
}
\value{
named character - the requested patterns with respective names
}
\description{
Get the file patterns for a given group. If the group does not exist, it will return NULL
}
\examples{
\dontrun{
get.pattern.group(group = "sas_downloads")
}

}
