% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/patterns.R
\name{get.pattern.info}
\alias{get.pattern.info}
\title{Get File Pattern Info}
\usage{
get.pattern.info(name)
}
\arguments{
\item{name}{character - name of pattern}
}
\value{
data.frame - the requested pattern info/fields (1 row)
}
\description{
Get the file pattern info (all fields) for a given name. If the name does not exist, it will return NULL
}
\examples{
\dontrun{
get.pattern(name = "raw_data_folder")
}

}
