% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/patterns.R
\name{get.patterns}
\alias{get.patterns}
\title{Get File Pattern Data
#'}
\usage{
get.patterns(names = ".*")
}
\arguments{
\item{names}{character pattern (see grep) to apply to names}
}
\value{
data frame with all pattern data
}
\description{
Get File Pattern Data
#'
}
