% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/table_output.R
\name{get_brfss_style_table}
\alias{get_brfss_style_table}
\title{Print BRFSS Data in CDC Format}
\usage{
get_brfss_style_table(
  coi,
  binary,
  num_vals,
  den_vals,
  exclude = c("Don.*t|Refuse"),
  weighted = TRUE,
  min_num = 0,
  subsets,
  ques = "",
  ...
)
}
\arguments{
\item{coi}{[character] column of interest}

\item{binary}{[logical] is this a simple yes/no}

\item{num_vals}{[integer] if binary=T, values representing the numerator}

\item{den_vals}{[integer] if binary=T, values representing the denominator}

\item{exclude}{[integer] if binary=F, values in the coi to exclude (usually representing [Don't know/Not sure] and [Refused])}

\item{weighted}{[character] column containing weight value}

\item{min_num}{[integer] minimun number in denominator for printing}

\item{subsets}{[character] columns to subset data by}

\item{ques}{[character] text of the survey question}

\item{...}{arguments to pass to other functions}

\item{df}{[data.frame] BRFSS data}
}
\description{
Print summary of results for a column using the style reported by CDC in their final report.
}
\details{
For example ... column CVDSTRK3

 (Ever told) you had a stroke. (CVDSTRK3)
                                    Yes                      No
        GROUP         TOTAL    N     %       C.I.      N     %       C.I.
 ===========================================================================
        Total           841    19    2.0   0.8-3.2    822   98.0  96.8-99.2
 ...........................................................................
         Male           308     8    1.7   0.3-3.0    300   98.3  97.0-99.7
        Female          533    11    2.3   0.5-4.1    522   97.7  95.9-99.5
 ...........................................................................
 White only, non-Hisp    98     1    0.6   0.0-1.7     97   99.4  98.3-100.0
 Black only, non-Hisp   558    10    0.9   0.3-1.6    548   99.1  98.4-99.7
 Asian only, non-Hisp    13     0    0.0   0.0-0.0     13  100.0  100.0-100.
 Multiracial, non-His    12     0    0.0   0.0-0.0     12  100.0  100.0-100.
       Hispanic         110     6    6.9   0.8-13.0   104   93.1  87.0-99.2
 Don’t know/Not sure/    33     1    2.0   0.0-5.9     32   98.0  94.1-100.0
 ...........................................................................
        18-24            39     0    0.0   0.0-0.0     39  100.0  100.0-100.
        25-34            75     2    3.9   0.0-9.2     73   96.1  90.8-100.0
        35-44            64     0    0.0   0.0-0.0     64  100.0  100.0-100.
        45-54           137     1    1.1   0.0-3.1    136   98.9  96.9-100.0
        55-64           203     6    3.1   0.4-5.8    197   96.9  94.2-99.6
        65+             299    10    3.3   1.0-5.6    289   96.7  94.4-99.0
 Don’t know/Refused/M    24     0    0.0   0.0-0.0     24  100.0  100.0-100.
 ...........................................................................
    Less than H.S.      148     7    4.4   0.7-8.0    141   95.6  92.0-99.3
     H.S. or GED        264     7    1.2   0.2-2.2    257   98.8  97.8-99.8
     Some College       155     0    0.0   0.0-0.0    155  100.0  100.0-100.
   College Graduate     253     5    1.7   0.1-3.3    248   98.3  96.7-99.9
 Don’t know/Not sure/    21     0    0.0   0.0-0.0     21  100.0  100.0-100.
 ...........................................................................
  Less than $14,999     136     2    1.6   0.0-4.2    134   98.4  95.8-100.0
  $15,000 - $24,999     159     4    2.6   0.0-5.8    155   97.4  94.2-100.0
  $25,000 - $34,999     102     4    2.0   0.0-4.2     98   98.0  95.8-100.0
  $35,000 - $49,999      96     2    1.1   0.0-2.7     94   98.9  97.3-100.0
  $50,000 - $74,999     106     0    0.0   0.0-0.0    106  100.0  100.0-100.
       $75,000+         146     2    0.8   0.0-1.8    144   99.2  98.2-100.0
 Don’t know/Not sure     35     2    6.7   0.0-16.9    33   93.3  83.1-100.0
       Refused           57     3    4.2   0.0-9.5     54   95.8  90.5-100.0
 ...........................................................................
}
