% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/layout.R
\name{get_fw_layout}
\alias{get_fw_layout}
\title{Get BRFSS Fixed Width Layout}
\usage{
get_fw_layout(year = NULL)
}
\arguments{
\item{year}{integer - year of interest}
}
\value{
data frame- 2 variables
}
\description{
Gets a data frame with information about the column names and the column widths for parsing a
BRFSS fixed width ASCII file.
}
\examples{
\dontrun{
get_fw_layout(2020)
}
}
