% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geogs.R
\name{get_geogs_all}
\alias{get_geogs_all}
\title{Get All Geographies}
\usage{
get_geogs_all()
}
\value{
data frame
}
\description{
Get the data frame containing all geographies
}
\examples{
\dontrun{
get_geogs_all()
}
}
