% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/table_output.R
\name{get_vertical_table}
\alias{get_vertical_table}
\title{Get Vertical Table}
\usage{
get_vertical_table(
  coi,
  binary = FALSE,
  num_vals,
  den_vals,
  exclude = c("Don.*t|Refuse"),
  weighted,
  min_num = 0,
  subsets,
  ques = "",
  ...
)
}
\arguments{
\item{coi}{character - column of interest}

\item{binary}{logical - only Yes/No values}

\item{num_vals}{integer: values from column in the numerator}

\item{den_vals}{integer: values from column in the denominator}

\item{subsets}{character: names of column to subset data by}

\item{...}{}

\item{df}{data.frame data of interest}
}
\description{
Get Vertical Table
}
