% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/report_funcs.R
\name{getMeasureReport}
\alias{getMeasureReport}
\title{BRFSS data file processing}
\usage{
getMeasureReport(df, meas, title = "Var1", order = F, useNA = "ifany",
  ...)
}
\arguments{
\item{df}{data.frame containing the measure of interest}

\item{meas}{character column name}

\item{title}{character}

\item{order}{boolean}

\item{useNA}{character}

\item{...}{parameters to be passed to other functions}
}
\value{
data.frame containing measure data
}
\description{
Reads the file containing layout data for a BRFSS raw data file
}

