% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/report_funcs.R
\name{getMeasureReportBreaks}
\alias{getMeasureReportBreaks}
\title{BRFSS data file processing}
\usage{
getMeasureReportBreaks(
  df,
  meas,
  breaks,
  title = "Desc",
  order = F,
  useNA = "ifany",
  debug = F,
  replace,
  replace_vals,
  ...
)
}
\arguments{
\item{df}{data.frame containing the measure of interest}

\item{meas}{character breaks}

\item{title}{character}

\item{order}{logical}

\item{useNA}{character}

\item{debug}{logical}

\item{replace}{character vector}

\item{replace_vals}{character vector}

\item{...}{parameters to be passed to other functions}

\item{format}{character}
}
\description{
Reads the file containing layout data for a BRFSS raw data file
}
