% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/report_funcs.R
\name{getMeasureReportEach}
\alias{getMeasureReportEach}
\title{BRFSS data file processing}
\usage{
getMeasureReportEach(
  df,
  meas,
  alias,
  title = "Desc",
  order = F,
  useNA = "ifany",
  debug = F,
  replace,
  replace_vals,
  format = "\%d",
  ...
)
}
\arguments{
\item{df}{data.frame containing the measure of interest}

\item{meas}{character column name}

\item{title}{character}

\item{order}{logical}

\item{useNA}{character}

\item{debug}{logical}

\item{replace}{character vector}

\item{replace_vals}{character vector}

\item{format}{character}

\item{...}{parameters to be passed to other functions}
}
\value{
data.frame containing measure data
#'
}
\description{
Reads the file containing layout data for a BRFSS raw data file
}
