% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/output.R
\name{horz_stats_simple}
\alias{horz_stats_simple}
\title{Simple horizontal table of summary stats for column}
\usage{
horz_stats_simple(df, coi, subsets, min_num = 0, percent = T, header = T,
  linesep, headsep, cat = F, invisible = F, html = F, tbltag = T,
  tbltagx = T, background, ...)
}
\arguments{
\item{df}{- data.frame}

\item{coi}{- character}

\item{subsets}{- character}

\item{min_num}{- integer}

\item{percent}{- logical}

\item{header}{- logical}

\item{linesep}{- character}

\item{headsep}{- character}

\item{cat}{- logical}

\item{invisible}{- logical}

\item{html}{- logical - output as HTML table}

\item{...}{- other parameters}
}
\value{
- character
}
\description{
Simple horizontal table of summary stats for column
}
\examples{

}

