% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/layout.R
\name{is_core}
\alias{is_core}
\title{Core Column?}
\usage{
is_core(year = NULL, col)
}
\arguments{
\item{year}{integer: year of interest}

\item{col}{character: column of interest}
}
\value{
logical
}
\description{
Core Column?
}
