% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/layout.R
\name{is_module}
\alias{is_module}
\title{Module Column?}
\usage{
is_module(year = NULL, col)
}
\arguments{
\item{year}{integer: year of interest}

\item{col}{character: column of interest}
}
\value{
logical
}
\description{
Module Column?
}
