% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/values.R
\name{make_factors}
\alias{make_factors}
\title{Convert BRFSS Columns to Factor}
\usage{
make_factors(df_brfss = NULL, df_layout = NULL, df_vals = NULL)
}
\arguments{
\item{df_brfss}{data.frame BRFSS data file}

\item{df_layout}{- layout data for the year}

\item{df_vals}{- value data for the year}

\item{year}{integer - year of interest}
}
\value{
data.frame of BRFSS data with factored columns
}
\description{
Use values file (created from codebook) to convert applicable BRFSS columns to factors
}
\examples{
\dontrun{
year <- 2020
df_brfss <- brfss_data(year = year, geog = "MT")
df_layout <- get.codebook.layout(year = year)
df_vals <- codebook_values(year = year)
df_brfss <- make_factors(df_brfss = df_brfss, df_layout = df_layout, df_vals = df_vals)
}

}
