% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/saq.R
\name{merge_saq_layout}
\alias{merge_saq_layout}
\title{Merge State Specific Questions with Standard Layout}
\usage{
merge_saq_layout(
  year = NULL,
  geog = NULL,
  df_layout = NULL,
  df_saq_layout = NULL
)
}
\arguments{
\item{df_saq_layout}{}
}
\value{

}
\description{
Merge State Specific Questions with Standard Layout
}
