% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modules.R
\name{module_data}
\alias{module_data}
\title{Module Data}
\usage{
module_data(year)
}
\arguments{
\item{year}{integer: year of interest}
}
\value{
data frame: modules by state
}
\description{
Module Data
}
\examples{
\dontrun{
df <- module_data(2021)
}
}
