% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modules.R
\name{module_geogs}
\alias{module_geogs}
\title{Get States Using BRFSS Module(s)}
\usage{
module_geogs(year, modules, versions = FALSE, reduce = TRUE)
}
\arguments{
\item{year}{- integer - 4-digit year of interest}

\item{modules}{- character or integer vector - modules of interest - can be module numbers or module names}

\item{versions}{- logical - include the version number}

\item{reduce}{- logical - reduce to a vector if appropriate}
}
\value{

}
\description{
Returns  a data frame that has the year, the version (if requested), the geog (2-char), and the module used or returns a
vector of modules if appropriate
}
\examples{
\dontrun{
module_geogs(2018,"Pre-Diabetes")
module_geogs(2018,c(1,2))
}

}
