% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/my_brfss.R
\name{my.brfss}
\alias{my.brfss}
\title{BRFSS Geography/Years of Interest}
\usage{
my.brfss(year = NULL, geog = NULL, other_geogs = NULL)
}
\arguments{
\item{year}{int vector of year(s) to set as default}

\item{geog}{character vector of geography abbreviations to set as default}

\item{other_geogs}{character vector of other geography abbreviations to process}
}
\value{
list with year, geography and other geographies (year, geog, other_geogs)
}
\description{
Set the default geography and year for working with BRFSS data
}
\examples{
\dontrun{
my.brfss(geog="MT", year=2021)
my.brfss(geog="MT", other_geogs)

}

}
