% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/my_brfss.R
\name{my.extent}
\alias{my.extent}
\title{BRFSS Data Source Extent}
\usage{
my.extent()
}
\value{
character data extent ('local' or 'public')
}
\description{
Get the extent for the BRFSS data of interest. The extent is either 'local' or 'public'. Generally, data
downloaded from the publicly available data set are 'public' since they contains data from all geographies,
and data downloaded from the secure CDC state-specific site are 'local' since they only contains data for that
geography (state or territory)
}
\examples{
\dontrun{
my.extent()

}

}
