% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/my_brfss.R
\name{my.other.geogs}
\alias{my.other.geogs}
\title{BRFSS Other Geographies of Interest}
\usage{
my.other.geogs()
}
\value{
character vector of default geometry(s)
}
\description{
Get the other geographies for working with BRFSS data. A region might be of interest,
rather than a single state. So other geographies might be c("MT","WY", "ID", "ND", "SD")
}
\examples{
\dontrun{
my.geog("MT")

}

}
