% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/my_brfss.R
\name{my.source}
\alias{my.source}
\title{BRFSS Data Source}
\usage{
my.source()
}
\value{
character data source ('sas' or 'ascii')
}
\description{
Get the source for the BRFSS data of interest
}
\examples{
\dontrun{
my.source()

}

}
