% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{my_brfss}
\alias{my_brfss}
\title{Collection of BRFSS States/Territories}
\format{
A list with 3 members
\describe{

  \item{year}{BRFSS year of interest}
  \item{geog}{primary geography (state/territory) of interest}
  \item{other_geogs}{other geographies (states/territories) of interest}
  ...
}
}
\usage{
my_brfss
}
\description{
A data frame containing  information for
}
\keyword{datasets}
