% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/patterns.R
\name{pattern.names}
\alias{pattern.names}
\title{File Pattern Names}
\usage{
pattern.names(filter = ".*")
}
\arguments{
\item{filter}{pattern (see grep) to apply to names}
}
\value{
character vector of pattern names
}
\description{
File Pattern Names
}
\examples{
\dontrun{
# get file pattern names that have the word folder in them
pattern.names(filter = "folder")
}

}
