% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/patterns.R
\name{pattern.requirements}
\alias{pattern.requirements}
\title{Pattern Requirements}
\usage{
pattern.requirements(pattern = ".*")
}
\arguments{
\item{pattern}{}
}
\value{
data frame - pattern name and arguments required for the named patterns
}
\description{
Get the parameter requirements for naming pattern(s).
}
\examples{
\dontrun{
pattern.requirements("codebook")
}

}
