% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/patterns.R
\name{patterns.local}
\alias{patterns.local}
\title{Set My Patterns to Local Default}
\usage{
patterns.local()
}
\description{
These naming patterns are generally used for data directly downloaded from the CDC passwor-protected data website
(https://nccd.cdc.gov/BRFSSStates/Index.html#/actions)
}
