% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/patterns.R
\name{patterns.national}
\alias{patterns.national}
\title{Set My Patterns to National Default}
\usage{
patterns.national()
}
\description{
These naming patterns are generally used for data directly downloaded from the CDC BRFSS Public data website
(https://www.cdc.gov/brfss/annual_data/annual_YYYY.html)
}
