% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/output.R
\name{prep_stats}
\alias{prep_stats}
\title{Prepare the stats for printing}
\usage{
prep_stats(df, coi, subsets, min_num = 0, percent = T, df_fields)
}
\arguments{
\item{df}{- data.frame}

\item{coi}{- character}

\item{subsets}{- character}

\item{min_num}{- integer}

\item{percent}{- logical}

\item{df_fields}{- data.frame}
}
\value{
- data.frame
}
\description{
Reformats
}
\examples{

}

