% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/codebook.R
\name{process_codebook}
\alias{process_codebook}
\title{Download Codebook and Extract the Layout}
\usage{
process_codebook(year = NULL, geog = NULL)
}
\arguments{
\item{year}{integer - year of interest}
}
\description{
Download the CDC prepared codebook for the BRFSS data files and extract the layout. The format of the URL is not consistent by year so more than one pattern is attempted and all should fail but one for a given year.
}
\examples{
\dontrun{
process_codebook(2020)
}
}
