% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/values.R
\name{quest_types}
\alias{quest_types}
\title{Calculates the Type of BRFSS Questions}
\usage{
quest_types(df_layout = NULL, df_vals = NULL)
}
\arguments{
\item{df_layout}{data frame containng the column layout derived from the codebook}

\item{df_vals}{data frame containng the column values derived from the codebook}
}
\value{
data frame continaing the type for each column
}
\description{
This is normally not called by the user, but used internally to get the factors for each question.
Values returned are: factor, range, or NA. Weighting variables are among the NA values.
}
