% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ascii_data.R
\name{read.ascii}
\alias{read.ascii}
\title{Convert a BRFSS raw ASCII data file}
\usage{
read.ascii(filename = NULL, layout = NULL, completes = T, verbose = FALSE)
}
\arguments{
\item{filename}{character name of the raw ascii data file}

\item{layout}{character name of the file containing the layout or data.frame containing the layout}

\item{completes}{logical whether or not to include only complete interviews (default=TRUE)}

\item{verbose}{logical print some extra information during processing (default = FALSE)}
}
\value{
data frame containing survey data
}
\description{
Read BRFSS raw ASCII data file and save as data frame in .RData file
}
