% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read.brfss.R
\name{read.brfss.all}
\alias{read.brfss.all}
\title{Read BRFSS raw landline and cell phone data file}
\usage{
read.brfss.all(year, month, index = 1, ...)
}
\arguments{
\item{year}{integer year of the data included in the file}

\item{month}{integer month[s] of the data included in the file}

\item{index}{integer index of the raw data file}

\item{...}{other parameters to pass to called functions}
}
\value{
data frame containing landline data
}
\description{
Read BRFSS raw data files containing cell phone data only. A range of months may be included.
}
\examples{
\dontrun{
df<-read.brfss.all(2016,2:5) # will read Feb-May data for 2016
}

}
