% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read.brfss.R
\name{read.brfss.ascii}
\alias{read.brfss.ascii}
\title{Read BRFSS raw ASCII data file}
\usage{
read.brfss.ascii(
  filename,
  layout,
  state = c("all", "my", "other"),
  otherstate = "*",
  completes = T
)
}
\arguments{
\item{filename}{character name of the raw ascii data file}

\item{layout}{character name of the file containing the layout or data.frame containing the layout}

\item{state}{character include data for "all" states (default), "my" state, or "other" states}

\item{completes}{boolean whether or not to inlcude only complete interviews (default=TRUE)}
}
\value{
data frame containing survey data
}
\description{
Read BRFSS raw ASCII data file
}
\examples{
\dontrun{
df<-read.brfss(filename="./data_raw/2016/05_MAY/VIMAY161.dat",
layout="./data_raw/var_layouts/layout_2016")
}

}
