% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read.brfss.R
\name{read.brfss.field_values}
\alias{read.brfss.field_values}
\title{BRFSS data file processing}
\usage{
read.brfss.field_values(year)
}
\arguments{
\item{year}{integer year of the data included in the file}
}
\value{
character default path/filename
}
\description{
Reads the file containing field values data for a BRFSS raw data file.
For example, the question for, "How many different types of cancer have you had? (CNCRDIFF)", the file contains
a record for each of the following: \cr \cr
CNCRDIFF,1,	Only one \cr
CNCRDIFF,2,	Two \cr
CNCRDIFF,3,	Three or more \cr
CNCRDIFF,7,	Don't know/Not Sure \cr
CNCRDIFF,9,	Refused \cr
}
