% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read.brfss.R
\name{read.brfss.landline}
\alias{read.brfss.landline}
\title{Read BRFSS raw landline data file}
\usage{
read.brfss.landline(year, month, index = 1, ...)
}
\arguments{
\item{year}{integer year of the data included in the file}

\item{month}{integer month of the data included in the file}

\item{index}{integer index of the raw data file}

\item{...}{other arguments to be passed to read.brfss.ascii}
}
\value{
data frame containing landline data
}
\description{
Read a BRFSS raw data file containing landline data only
}
\examples{
\dontrun{
df<-read.brfss.landline(2016,2)
}

}
