% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read.brfss.R
\name{read.brfss.layout}
\alias{read.brfss.layout}
\title{BRFSS data file processing}
\usage{
read.brfss.layout(filename, year)
}
\arguments{
\item{year}{integer year of the data included in the file}
}
\value{
character default path/filename
}
\description{
Reads the file containing layout data for a BRFSS raw data file
}
