% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sas.R
\name{read.sas.field.ranges}
\alias{read.sas.field.ranges}
\title{Read Data Element Ranges from SASOUT file}
\usage{
read.sas.field.ranges(year)
}
\arguments{
\item{year}{integer: year of interest}
}
\value{
data.frame: the ranges
}
\description{
These ranges can be used to read the ascii fixed width file fprovided by CDC
}
\examples{
\dontrun{
read.sas.field.ranges(2021)

}
}
