% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sas.R
\name{read.xpt}
\alias{read.xpt}
\title{Read BRFSS Data from SAS XPT File}
\usage{
read.xpt(version = 0, verbose = F)
}
\arguments{
\item{version}{integer: version of interest (default is 0 ... main survey)}

\item{verbose}{logical: provide extra information during processing}
}
\description{
Creates a data frame from the xpt file provided by CDC. The sasout file is used to provide
some useful attributes for each column.
This function gets the default file names folder location from the naming_patterns data.
}
\examples{
\dontrun{
df_2021 <- read.xpt(2021)

df_data <- read.xpt(2020, version= 1)
}
}
