% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/codebook.R
\name{read_codebook}
\alias{read_codebook}
\title{Read the CDC Provided Codebook File}
\usage{
read_codebook(file = NULL, year = NULL)
}
\arguments{
\item{file}{character name of layout file}
}
\value{
character vector containing the lines in the file
}
\description{
If a filename is not provided, it uses the file_pattern data.frame
to figure out the name of the file. The layout data.frame is used to
parse the BRFSS ascii fixed width file, and to populate column attributes with information such
as label, survey section module number. If a file name is not passed, then my.brfss() is called to
determine the year and geographical place of interest. The geographical place of interest
may not be used, but will be passed for pattern matching of file name in case it is needed
}
