% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sas.R
\name{read_sas_field_ranges}
\alias{read_sas_field_ranges}
\title{Read Data Element Ranges from SASOUT file}
\usage{
read_sas_field_ranges()
}
\arguments{
\item{year}{integer: year of interest}
}
\value{
data.frame: the ranges
}
\description{
These ranges can be used to read the ascii fixed width file fprovided by CDC
}
\examples{
\dontrun{
read_sas_field_ranges(2021)

}
}
