% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sas.R
\name{read_sasout}
\alias{read_sasout}
\title{Read BRFSS SASOUT file produced by CDC}
\usage{
read_sasout()
}
\arguments{
\item{year}{integer: year of interest}

\item{folder}{character: folder location}

\item{file}{character: file name}

\item{version}{integer: version number (defaults = 0)}
}
\value{
data frame
}
\description{
Read BRFSS SASOUT file produced by CDC
}
