% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/patterns.R
\name{relocate_by_pattern}
\alias{relocate_by_pattern}
\title{Move File to Standard Location}
\usage{
relocate_by_pattern(name, file, ...)
}
\arguments{
\item{name}{character pattern name}

\item{file}{character file to be renamed}

\item{...}{value pairs to be passed to the apply.pattern function}
}
\value{
logical status of operation
}
\description{
Move a file (using file.rename()) to aa location specified in the standard (expected)
 location using pattern naming.
}
\examples{
\dontrun{
relocate_by_pattern("ascii_path_raw", "./temp/ascii.dat", YEAR = 2021, VERS = 0)
}
}
