% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/versions.R
\name{responses_by_geog}
\alias{responses_by_geog}
\title{Response Totals}
\usage{
responses_by_geog()
}
\arguments{
\item{year}{integer: the 4-digit year of interest}

\item{geog}{character or integer: either the 2-char abbrev, the full name,
or the FIPS ID for the geography of interest}
}
\value{
data frame with geog version and count
}
\description{
Response Totals
}
