% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/saq.R
\name{saq_values}
\alias{saq_values}
\title{Get State-Added-Question Values}
\usage{
saq_values()
}
\value{
data frame
}
\description{
Get State-Added-Question Values
}
\examples{
\dontrun{
df_saq_values <- saq_values(2021)
}
}
