% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sas.R
\name{sas.download.xpt}
\alias{sas.download.xpt}
\title{Download SAS XPT files from CDC website}
\usage{
sas.download.xpt(year = NULL)
}
\arguments{
\item{year}{integer - year of interest}
}
\description{
Download SAS XPT files from CDC website
}
