% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sas.R
\name{sas_download_xpt}
\alias{sas_download_xpt}
\title{Download SAS XPT files from CDC website}
\usage{
sas_download_xpt(folderout = NULL, year = NULL, progress = NULL)
}
\arguments{
\item{year}{integer - year of interest}

\item{...}{other arguments to be passed to apply.pattern()}
}
\description{
Download SAS XPT files from CDC website
}
