% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sas.R
\name{sas_process_year}
\alias{sas_process_year}
\title{Process BRFSS Annual Data Files
Use this function to process a single year of BRFSS data. This function sill download, unzip,
and create a data frame with all data from the XPT file, as well as any other versions of the survey.}
\usage{
sas_process_year(
  year = NULL,
  download = TRUE,
  convert = TRUE,
  codebook = TRUE,
  split = TRUE,
  verbose = FALSE,
  ...
)
}
\arguments{
\item{year}{- int - year of interest}

\item{download}{- logical - download the data? Useful (set = FALSE) if you already have the downloaded files}

\item{convert}{- logical - read the xpt files into data_frames and save? Useful (set = FALSE) if you already have the xpt files processed}

\item{codebook}{- logical - download the annual codebook}

\item{split}{- logical - split the xpt file by state/geography}

\item{...}{other params}
}
\value{
invisible()
}
\description{
Process BRFSS Annual Data Files
Use this function to process a single year of BRFSS data. This function sill download, unzip,
and create a data frame with all data from the XPT file, as well as any other versions of the survey.
}
