% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process.R
\name{save_brfss}
\alias{save_brfss}
\title{Save BRFSS Data}
\usage{
save_brfss(df, year = NULL, geog = NULL, version = 0)
}
\arguments{
\item{df}{data frame - BRFSS data}

\item{year}{integer - year of interest}

\item{geog}{character - 2-char state/territory abbreviation}

\item{version}{integer - version of interest (default = 0)}
}
\description{
Save BRFSS data frame to a .RData file. This standardizes the object name and the file name
}
\examples{
\dontrun{
save_brfss <- function(df_brfss, year= 202, geog = "MT")
}
}
