% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/codebook.R
\name{save_codebook_layout}
\alias{save_codebook_layout}
\title{Build the Layout from CDC Provided Codebook}
\usage{
save_codebook_layout(file = NULL, year = NULL, ...)
}
\arguments{
\item{file}{character name of layout file}

\item{...}{other arguments to be passed to apply.pattern()}
}
\value{
data.frame containing the layout to be used for decoding/parsing the ascii file
}
\description{
If a filename is not provided, it uses the file_pattern data.frame
to figure out the name of the file. The layout data.frame is used to
parse the BRFSS ascii fixed width file, and to populate column attributes with information such
as label, survey section module number. If a file name is not passed, then my.brfss() is called to
determine the year and geographical place of interest. The geographical place of interest
may not be used, but will be passed for pattern matching of file name in case it is needed
}
