% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/versions.R
\name{save_response_stats}
\alias{save_response_stats}
\title{Save Response Statistics}
\usage{
save_response_stats()
}
\arguments{
\item{year}{integer - year of interest}

\item{...}{other parameters}
}
\description{
Calculate the number of responses for each version by geography
}
\examples{
\dontrun{
save_response_stats(2018)
}
}
