% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/layout.R
\name{section_type}
\alias{section_type}
\title{Section of BRFSS Column}
\usage{
section_type(year = NULL, col)
}
\arguments{
\item{year}{integer: year of interest}

\item{col}{character: column of interest}
}
\value{
character
}
\description{
Gets the section for a BRFSS column from the layout. Useful for figuring out if a
 column of interest is from the core or from a module.
}
