% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/patterns.R
\name{set.pattern}
\alias{set.pattern}
\title{Set Naming Pattern}
\usage{
set.pattern(name, pattern = NULL, group = "", desc = "")
}
\arguments{
\item{name}{character: name of pattern}

\item{group}{}
}
\value{
Nothing
}
\description{
If the name exists already, it will overwrite (update) that name without warning.
If the name does not exist, it will add it.
}
\examples{
\dontrun{
set.pattern(name = "raw_data_folder", pattern = "./data_raw/", desc = "Folder for raw BRFSS data")
}

}
