% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/codebook.R
\name{set_codebook_ext}
\alias{set_codebook_ext}
\title{Set Codebook Extension}
\usage{
set_codebook_ext(ext)
}
\arguments{
\item{ext}{character - extension (pdf, rtf, txt, or html)}
}
\description{
Convenience function to set the codebook extension
}
\examples{
\dontrun{
set_codebook_ext("pdf")
}

}
