% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/values.R
\name{simple_counts}
\alias{simple_counts}
\title{Simple Response Counts}
\usage{
simple_counts(
  df_brfss,
  coi,
  na.rm = TRUE,
  valid = FALSE,
  total = TRUE,
  pct = TRUE
)
}
\arguments{
\item{df_brfss}{data frame - BRFSS data}

\item{coi}{character or symbol - the column to report on}

\item{na.rm}{logical - remove NAs (default = TRUE)}

\item{valid}{logical - remove "Don't Know" and "Refused" responses (default = FALSE)}

\item{total}{logical - show a total line (default = TRUE)}

\item{pct}{logical - include a percent column (default = TRUE)}
}
\description{
Report counts of responses in a simply formatted table. Options include removing NAs, including only valid responses,
including totals, and including percents. Table is printed to console.
}
\examples{
\dontrun{
df_brfss <- brfss_data(year = 2021, geog = "VI")
df_brfss \%>\% simple_counts(COVIDVAC, valid = T)
}
}
