% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/survey.R
\name{simple_stats}
\alias{simple_stats}
\title{Simple Percents and Confidence Intervals}
\usage{
simple_stats(
  df_brfss = NULL,
  year = NULL,
  geog = NULL,
  coi,
  wt = "_llcpwt",
  strata = "_ststr",
  exclude = c("Don.t|Refuse")
)
}
\arguments{
\item{df_brfss}{data.frame: the data frame with the columns nedded}

\item{year}{integer: year of interest}

\item{geog}{character: geography of interest}

\item{coi}{character: column of interest}

\item{wt}{character: column name of weights}

\item{strata}{character: column name of strata}

\item{exclude}{character: pattern for excludes}
}
\value{
Object of class "svystat"
}
\description{
Get the percent and CI for a column of data
}
