% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process.R
\name{split_geogs}
\alias{split_geogs}
\title{Split BRFSS Data by Geography}
\usage{
split_geogs(
  geogs = NULL,
  main = TRUE,
  versions = TRUE,
  col_atts = TRUE,
  factorize = TRUE,
  verbose = TRUE,
  progress = NULL
)
}
\arguments{
\item{main}{logical - process main XPT file}

\item{versions}{logical - process all versioned XPT/ASC files}

\item{verbose}{logical - provide details during processing}

\item{my_geog}{character - abbreviation for primary state/geography of interest (e. "MT")}

\item{other_geogs}{character - abbreviations for other states/geographies of interest (e. c("ID","WY"))}
}
\description{
The main BRFSS data file that is created when the downloaded file from the public CDC website
is read has data for all geographies, whether the SAS (XPT) file or the ASCII (ASC) file.
This function splits out the geographies of interest.
}
\examples{

\dontrun{
split_geogs( main=TRUE, versions=TRUE,verbose=TRUE)
}


}
