% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modules.R
\name{state_modules}
\alias{state_modules}
\title{Get BRFSS Modules Used by State(s)}
\usage{
state_modules(year, states, versions = FALSE, reduce = TRUE)
}
\arguments{
\item{year}{- integer - 4-digit year of interest}

\item{states}{- character or integer vector - states of interest - can be 2 character state code or FIPS code (integer)}

\item{versions}{- logical - include the version number}

\item{reduce}{- logical - reduce to a vector if appropriate}
}
\value{

}
\description{
Returns  a data frame that has the year, the version (if requested), the state (2-char), and the module used or returns a
vector of modules if appropriate
}
\examples{
\dontrun{
state_modules(2018,"WI")
state_modules(2018,c(1,2))
state_modules(2018,"MI",versions=T)
}

}
