% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/survey.R
\name{survey_stats_binary}
\alias{survey_stats_binary}
\title{Survey Statistics}
\usage{
survey_stats_binary(df0, coi, num_vals, den_vals, ...)
}
\arguments{
\item{df0}{- data.frame: the survey data}

\item{coi}{- character: column name of interest}

\item{num_vals}{- integer: values from column in the numerator}

\item{den_vals}{- integer: values from column in the denominator}

\item{year}{- integer: if missing will calculate from df}

\item{subset}{- character: name of column to subset data by}

\item{conf}{- numeric: confidence level, default is .95}

\item{weighting}{- character: name of column with weighting}
}
\value{
data.frame: statistics for each subset
}
\description{
Retrieve survey statistics from survey data set for a specific column with optional subsetting by other column(s) and
optional weighting. This coerces the values in num_vals (numerator values) to 'Yes' and all values in den_vals (denominator values)
not in num_vals to 'No'
}
\details{
Retuns a data.frame containing columns identiying the calculated numerators, denominators, means, standard error, and
confidence interval (lower and upper) for either the entire column of interest, or subsets of that column
#'
}
\examples{


}

