% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/patterns.R
\name{try.patterns}
\alias{try.patterns}
\title{Try Patterns with Set of Arguments}
\usage{
try.patterns(names = ".*", ...)
}
\arguments{
\item{names}{character pattern to be applied to pattern names to subset result}

\item{...}{arguments to be passed to apply.pattern()}
}
\description{
View all patterns with a given set of arguments. Useful for pattern naming testing.
}
\examples{
\dontrun{
try.patterns("codebook", YEAR = 2018, GEOG = "MT", VERS = 0, EXT = 'local')
}

}
