% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/values.R
\name{valid_only}
\alias{valid_only}
\title{Subset Data For Valid Responses}
\usage{
valid_only(
  df,
  coi,
  other_cols = character(0),
  invalid = c("don.{0,1}t know|refused")
)
}
\arguments{
\item{df}{data frame - BRFSS data}

\item{coi}{character - column to use for filtering}

\item{other_cols}{- other columns to keep}

\item{invalid}{- pattern for invalid rsponses}
}
\value{
the subsetted data frame
}
\description{
Removes 'Don't Know' and 'Refused' as well as Missing responses
}
\details{
Invalid responses are indicated by the `invalid` pattern. Include other columns,
such as weighting and others for subsetting, like sex or age.

The default pattern for invalid responses is "don.{0,1}t know|refused" (regexp formatted)
}
\examples{
\dontrun{
df <- brfss_data()
coi <- "GENHLTH"
other_cols <- c("_LLCPWT","_SEX","_AGE65YR")
df_coi <- df \%>\% valid_only(coi = coi, other_cols = other_cols)
}

}
