% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/values.R
\name{values}
\alias{values}
\title{Get Values}
\usage{
values(year = NULL)
}
\arguments{
\item{year}{integer - year of interest for codebook}
}
\description{
The annual codebook is parsed for possible values for each variable and saved. If SAQ Values are available,
they are merged. This function retrieves those values
}
