% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/output.R
\name{vert_stats_simple}
\alias{vert_stats_simple}
\title{Title}
\usage{
vert_stats_simple(df, coi, subsets, min_num = 0, percent = T, cat = FALSE,
  invisible = FALSE, ...)
}
\arguments{
\item{df}{- data.frame}

\item{coi}{- character}

\item{subsets}{- character}

\item{min_num}{- integer}

\item{percent}{- logical}

\item{cat}{- logical}

\item{invisible}{- logical}

\item{...}{- other parameters}
}
\value{
- data.frame
}
\description{
Title
}
\examples{

}

