% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ascii_data.R
\name{ascii.download.data}
\alias{ascii.download.data}
\title{Download ASCII BREFSS data}
\usage{
ascii.download.data(year = NULL, destpath = NULL, unzip = TRUE, rmzip = TRUE)
}
\arguments{
\item{year}{integer - year of data file}

\item{destpath}{character - location to store file - will be created if it does not exist - default is the data_raw folder in the project root folder}

\item{unzip}{logical - unzip file?}

\item{rmzip}{logical - remove the zip file?}
}
\value{

}
\description{
Download the annual ASCII BRFSS data file
}
\examples{
\dontrun{
ascii.download.data(2016)
}

}
