% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/survey.R
\name{brfss_data}
\alias{brfss_data}
\title{Get BRFSS Survey Data}
\usage{
brfss_data(year = NULL, geog = NULL, version = 0, extent = "local")
}
\arguments{
\item{year}{- int - year of interest}

\item{geog}{- character - 2-char geog of interest (ex: "NY")}

\item{version}{- survey version; 0 (default) = main survey}
}
\value{
data frame
}
\description{
Get BRFSS Survey Data
}
\examples{
\dontrun{
df <- brfss_data(2020, geog = "MD", version = 1)

}

}
