% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/saq.R
\name{build_saq_values}
\alias{build_saq_values}
\title{Build State-Added-Questions Values File}
\usage{
build_saq_values(year = NULL, geog = NULL)
}
\arguments{
\item{year}{integer year of interest}

\item{geog}{character geog of interest (or none, the default geography)}
}
\description{
Read in .csv file and save as .rda file
}
\examples{
\dontrun{
build_saq_values(2021)
}
}
