% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/values.R
\name{codebook_values}
\alias{codebook_values}
\title{Get Codebook Values}
\usage{
codebook_values()
}
\arguments{
\item{year}{integer - year of interest for codebook}

\item{...}{other arguments to be passed to apply.pattern()}
}
\description{
The annual codebook is parsed for possible values for each variable and saved. This function retrieves those values
}
