% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/codebook.R
\name{get.codebook.layout}
\alias{get.codebook.layout}
\title{Get layout created from the codebook}
\usage{
get.codebook.layout()
}
\arguments{
\item{year}{integer - year of interest}
}
\value{
data frame with layout data
}
\description{
Get layout created from the codebook
}
