% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/patterns.R
\name{get.pattern}
\alias{get.pattern}
\title{Get File Pattern}
\usage{
get.pattern(name, expand = FALSE)
}
\arguments{
\item{name}{character - name of pattern}

\item{expand}{logical - convert sub patterns}
}
\value{
character - the requested pattern
}
\description{
Get the file pattern for a given name. If the name does not exist, it will return NULL
}
\examples{
\dontrun{
get.pattern(name = "raw_data_folder")
}

}
