% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/table_output.R
\name{get_vertical_table}
\alias{get_vertical_table}
\title{Title}
\usage{
get_vertical_table(
  df,
  coi,
  binary,
  num_vals,
  den_vals,
  exclude,
  weighting,
  min_num = 0,
  subsets,
  ques = "",
  ...
)
}
\arguments{
\item{df}{data.frame data of interest}

\item{coi}{character - column of interest}

\item{binary}{logical - only Yes/No}

\item{num_vals}{integer: values from column in the numerator}

\item{den_vals}{integer: values from column in the denominator}

\item{subsets}{character: names of column to subset data by}

\item{...}{}
}
\value{

}
\description{
Title
}
\examples{

}
