% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/table_output.R
\name{get_vertical_table}
\alias{get_vertical_table}
\title{Title}
\usage{
get_vertical_table(df, coi, binary, num_vals, den_vals, exclude, weighting,
  min_num = 0, subsets, ques = "", ...)
}
\arguments{
\item{...}{}
}
\description{
Title
}
\examples{

}

