% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/versions.R
\name{highest_version}
\alias{highest_version}
\title{Get Highest Version of Survey for a Year}
\usage{
highest_version(year = NULL, ...)
}
\arguments{
\item{year}{- the 4-digit year of interest}

\item{...}{other arguments to be passed to apply.pattern()}
}
\value{
integer - the highest version number found in the filenames
}
\description{
Looks through the filenames in the BRFSS data folder for files with the pattern  *_Vn.* where n is the version number.
}
\examples{
\dontrun{
maxvers<-highest_version(2018)
}

}
