% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modules.R
\name{modules_used}
\alias{modules_used}
\title{Modules Used by Geography}
\usage{
modules_used(year = NULL, extent = NULL, source = NULL, ...)
}
\arguments{
\item{year}{integer - year of interest}

\item{geogs}{character - geography(s) of interest}
}
\value{
data frame of module information
}
\description{
Modules Used by Geography
}
\examples{
\dontrun{
modules_used(2020, "MT")
}


}
