% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/my_brfss.R
\name{my.geog}
\alias{my.geog}
\title{BRFSS Geography of Interest}
\usage{
my.geog()
}
\value{
character vector of default geometry(s)
}
\description{
Get the default geographyfor working with BRFSS data
}
\examples{
\dontrun{
my.geog("MT")

}

}
