% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/patterns.R
\name{patternize}
\alias{patternize}
\title{Substitute Variables in Pattern}
\usage{
patternize(strIn, ...)
}
\arguments{
\item{strIn}{character - patterns to apply the variables to}

\item{...}{list - variables to insert in form of c(YEAR = xxxx, VERS = 1), etc}
}
\value{
character vector of modified strings
}
\description{
Substitute Variables in Pattern
}
\examples{
year <- 2020
geog <- "MT"
patternize("XPT_[GEOG]_[YEAR]", YEAR = year, GEOG = geog)

}
