% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process.R
\name{process_year}
\alias{process_year}
\title{Process BRFSS Annual Data Files}
\usage{
process_year(
  year = NULL,
  source = NULL,
  download = TRUE,
  layout = TRUE,
  convert = TRUE,
  codebook = TRUE,
  split = TRUE,
  factorize = TRUE,
  saq = FALSE,
  verbose = FALSE,
  geog = NULL,
  extent = NULL,
  ...
)
}
\arguments{
\item{year}{- int - year of interest}

\item{source}{character - data source ('sas' or 'ascii')}

\item{download}{- logical - download the data? Useful (set = FALSE) if you already have the downloaded files}

\item{convert}{- logical - read the downloaded files into data_frames and save?
Useful (set = FALSE) if you already have the downloaded files processed}

\item{codebook}{- logical - download and process the annual codebook}

\item{split}{- logical - split the processed data file by state/geography}

\item{factorize}{- logical - convert columns to factors where appropriate}

\item{...}{other params}
}
\value{
invisible()
}
\description{
Use this function to process a single year of BRFSS data. This function sill download, unzip,
and create a data frame with all data from the XPT file, as well as any other versions of the survey.
}
