% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process.R
\name{process_year}
\alias{process_year}
\title{Process BRFSS Annual Data Files}
\usage{
process_year(
  dl_metadata = FALSE,
  dl_codebook = FALSE,
  dl_data = FALSE,
  layout = TRUE,
  convert = TRUE,
  codebook = TRUE,
  attribs = TRUE,
  split = TRUE,
  factorize = TRUE,
  saq = FALSE,
  responses = TRUE,
  verbose = FALSE,
  progress = NULL,
  ...
)
}
\arguments{
\item{convert}{- logical - read the downloaded files into data_frames and save?
Useful (set = FALSE) if you already have the downloaded files processed}

\item{codebook}{- logical - download and process the annual codebook}

\item{attribs}{- logical - add attributes to each column describing the column's data}

\item{split}{- logical - split the processed data file by state/geography}

\item{factorize}{- logical - convert columns to factors where appropriate}

\item{saq}{- logical - add state-added questions}

\item{progress}{- function - progress function to pass progress info to}

\item{...}{other params}

\item{year}{- int - year of interest}

\item{source}{character - data source ('sas' or 'ascii')}

\item{download}{- logical - download the data? Useful (set = FALSE) if you already have the downloaded files}
}
\value{
invisible()
}
\description{
Use this function to process a single year of BRFSS data. This function sill download, unzip,
and create a data frame with all data from the XPT file, as well as any other versions of the survey.
}
