% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read.brfss.R
\name{read.brfss.ascii}
\alias{read.brfss.ascii}
\title{Read BRFSS raw ASCII data file}
\usage{
read.brfss.ascii(
  year = NULL,
  filename = NULL,
  layout = NULL,
  state = c("all", "my", "other"),
  otherstate = "*",
  completes = T
)
}
\arguments{
\item{year}{integer year of interest}

\item{filename}{character name of the raw ascii data file}

\item{layout}{character name of the file containing the layout or data.frame containing the layout}

\item{state}{character include data for "all" states (default), "my" state, or "other" states}

\item{completes}{logical whether or not to include only complete interviews (default=TRUE)}
}
\value{
data frame containing survey data
}
\description{
Read BRFSS raw ASCII data file
}
\examples{
\dontrun{
my.brfss(year = 2020, geog = "MT")
df<-read.brfss.ascii(year = ,
layout="./data_raw/var_layouts/layout_2016")
}

}
