% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/saq.R
\name{saq_layout}
\alias{saq_layout}
\title{Get State-Added-Questions Layout}
\usage{
saq_layout(year, geog = "")
}
\arguments{
\item{year}{integer year of interest}

\item{geog}{character geog of interest (or none, the default geography)}
}
\value{
data frame
}
\description{
Get State-Added-Questions Layout
}
\examples{
\dontrun{
df_saq_layout <- saq_layout(2021)
}
}
