% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sas.R
\name{sas_process_year}
\alias{sas_process_year}
\title{Process BRFSS Annual Data Files
Use this function to process a single year of BRFSS data. This function sill download, unzip,
and create a data frame with all data from the XPT file, as well as any other versions of the survey.}
\usage{
sas_process_year(
  dl_metadata = FALSE,
  dl_codebook = FALSE,
  dl_data = FALSE,
  layout = TRUE,
  codebook = TRUE,
  saq = FALSE,
  convert = TRUE,
  split = TRUE,
  responses = TRUE,
  factorize = TRUE,
  verbose = FALSE,
  progress = NULL,
  ...
)
}
\arguments{
\item{layout}{- logical - save sas layout}

\item{codebook}{- logical - download and process the annual codebook}

\item{saq}{- logical - are there state-added questions for processing}

\item{convert}{- logical - read the xpt files into data_frames and save? Useful (set = FALSE)
if you already have the xpt files processed}

\item{split}{- logical - split the xpt file by state/geography}

\item{factorize}{- logical - add values as factors to columns}

\item{...}{further arguments passed to other methods}

\item{year}{- int - year of interest}

\item{download}{- logical - download the data? Useful (set = FALSE) if you already have the downloaded files}

\item{geog}{- character - geography of interest}
}
\value{
invisible()
}
\description{
Process BRFSS Annual Data Files
Use this function to process a single year of BRFSS data. This function sill download, unzip,
and create a data frame with all data from the XPT file, as well as any other versions of the survey.
}
\examples{
\dontrun{
sas_process_year(year = 2020, download=TRUE, layout = TRUE, codebook = TRUE, convert = TRUE, split = TRUE)

}
}
