% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process.R
\name{split_geogs}
\alias{split_geogs}
\title{Split BRFSS Data by Geography}
\usage{
split_geogs(
  year = NULL,
  source = NULL,
  main = TRUE,
  versions = TRUE,
  my_geog = NULL,
  other_geogs = NULL,
  factorize = FALSE,
  verbose = TRUE
)
}
\arguments{
\item{year}{integer - year of interest}

\item{source}{character - data source ('sas' or 'ascii')}

\item{main}{logical - process main XPT file}

\item{versions}{logical - process versioned XPT file}

\item{my_geog}{character - abbreviation for primary state/geography of interest (e. "MT")}

\item{other_geogs}{character - abbreviations for other states/geographies of interest (e. c("ID","WY"))}

\item{verbose}{logical - provide details during processing}
}
\description{
The main BRFSS  data file created when the file is downloaded and read has data
for all geographies, whether the SAS (XPT) file or the ASCII (ASC) file. This function splits out the geographies of interest.
}
\examples{

\dontrun{
split_geogs(year = 2020, source = 'sas', main=TRUE, versions=TRUE, my_geog="MT", other_geogs=NULL,verbose=TRUE)
}


}
