% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stats.R
\name{survey_stats}
\alias{survey_stats}
\title{Survey Statistics}
\usage{
survey_stats(
  df_data = NULL,
  coi,
  exclude = c("Don.*t|Refuse"),
  subvars = NULL,
  subset_by = NULL,
  sub_exclude = c("Don.*t|Refuse"),
  conf = 0.95,
  weighted = TRUE,
  weight_col = "_LLCPWT",
  pct = FALSE,
  digits = 99
)
}
\arguments{
\item{df_data}{- data.frame: the survey data}

\item{coi}{- character: column of interest name for analysis}

\item{exclude}{- integer: values in coi to exclude for analysis purposes}

\item{conf}{- numeric: confidence level, default is .95}

\item{weighted}{- logical: use weighting}

\item{subvar}{- character: name of column to subset data by}
}
\value{
data.frame: statistics for the coi each subset
}
\description{
Retrieve survey statistics from survey data set for a specific column with optional subsetting by other column(s) and
optional weighting. This coerces the values in num_vals (numerator values) to 'Yes' and all values in den_vals (denominator values)
not in num_vals to 'No'
}
\details{
Retuns a data.frame containing columns identiying the calculated numerators, denominators, means, standard error, and
confidence interval (lower and upper) for either the entire column of interest, or subsets of that column
#'
}
